+++ src/App.css
@@ -12,7 +12,6 @@
     animation: App-logo-spin infinite 20s linear;
   }
 }
-
 .App-header {
   background-color: #282c34;
   min-height: 100vh;
@@ -36,8 +35,3 @@
     transform: rotate(360deg);
   }
 }
-
-ul li{
-  font-size:14px;
-  display:inline-block;
-  }
