--- src/Galeria.jsx
@@ -1,5 +1,0 @@
-function Galeria() {
-    return(<div>Галерея тута</div>)
-}
-
-export default Galeria;
