+++ src/Galeria/Galeria.jsx
@@ -1,0 +1,5 @@
+function Galeria() {
+  return <div>Галерея тута</div>;
+}
+
+export default Galeria;
