+++ src/Header/Header.css
@@ -1,0 +1,4 @@
+ul li {
+  font-size: 14px;
+  display: inline-block;
+}
